requirements.txt
absl-py==0.11.0
asgiref==3.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.15.1
dlib @ file:///D:/bld/dlib_1607165191327/work
dnspython==2.6.1
drf-yasg==1.21.7
flask==3.0.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==8.0.0
imutils==0.5.4
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
keras==2.15.0
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==2.1.5
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mtcnn==0.1.1
networkx==2.5
numpy==1.23.5
oauthlib==3.1.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.10.1
six @ file:///C:/ci/six_1605187374963/work
sqlparse==0.5.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.56.1
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.3
werkzeug==3.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.19.2